generalize extra models (#96)
