+++ backend/cms/models/extra.py
@@ -1,31 +1,25 @@
-from abc import abstractmethod
-
 from django.db import models
-
 from cms.models import Site
+from cms.models.extra_template import ExtraTemplate
 
 
 class Extra(models.Model):
     site = models.ForeignKey(Site)
+    template = models.ForeignKey(ExtraTemplate)
     created_date = models.DateTimeField(auto_now_add=True)
     last_updated = models.DateTimeField(auto_now=True)
 
-    @abstractmethod
     def alias(self):
-        pass
+        return self.template.alias
 
-    @abstractmethod
     def name(self):
-        pass
+        return self.template.name
 
-    @abstractmethod
     def thumbnail(self):
-        pass
+        return self.template.thumbnail
 
-    @abstractmethod
     def url(self):
-        pass
+        return self.template.url
 
-    @abstractmethod
     def post_data(self):
-        pass
+        return self.template.post_data
