+++ backend/cms/models/postal_code_in_get_parameter_extra.py
@@ -1,0 +1,7 @@
+from cms.models.extra import Extra
+
+
+class PostalCodeInGetParameterExtra(Extra):
+
+    def url(self):
+        return self.template.url + self.site.postal_code
