+++ backend/cms/models/postal_code_in_post_parameter_extra.py
@@ -1,0 +1,9 @@
+from cms.models.extra import Extra
+
+
+class PostalCodeInPostParameterExtra(Extra):
+
+    def post_data(self):
+        post = self.template.post_data
+        post.update({'search-plz': self.site.postal_code})
+        return post
